+++ src/elements/Button.vue
@@ -2,12 +2,13 @@
   <component
     :is="to ? 'a' : 'button'"
     @click="onClick"
-    :href="to || undefined"
+    :href="!disabled && !cardDisabled ? to : undefined"
     :type="to ? undefined : 'button'"
     :class="[
       // base class
       'btn',
       'text-xs sm:text-sm', // responsive text size
+      'transition-all', // transition effect
       { 'w-full': props.block }, // conditional full width
       disabled || cardDisabled
         ? 'bg-gray-100 cursor-not-allowed'
@@ -19,7 +20,7 @@
       computedBorderType, // border type class
       computedActiveColor, // active effect class
     ]"
-    :disabled="disabled || cardDisabled"
+    :disabled="disabled || cardDisabled ? true : undefined"
   >
     <!-- Loading Icon -->
     <span
@@ -96,7 +97,6 @@ const props = withDefaults(defineProps<ButtonProps>(), {
   block: false,
   outline: false,
   shadow: false,
-  disabled: false,
   borderType: "solid",
   isLoading: false,
   loadingIcon: "IconLoader",
@@ -118,6 +118,7 @@ const computedColor = computed(() => {
   if (props.color)
     if (props.outline) {
       const outlinecolors = {
+        default: "btn-outline-white",
         primary: "btn-outline-primary",
         info: "btn-outline-info",
         success: "btn-outline-success",
@@ -130,6 +131,7 @@ const computedColor = computed(() => {
       return outlinecolors[props.color];
     } else {
       const colors = {
+        default: "btn-white",
         primary: "btn-primary",
         info: "btn-info",
         success: "btn-success",
@@ -203,32 +205,34 @@ const onClick = () => {
 </script>
 
 <style scoped lang="scss">
+@use "sass:color";
+
 .btn-primary:active,
 .btn-outline-primary:active {
-  background-color: darken(#4361ee, 20%) !important;
+  background-color: color.adjust(#4361ee, $lightness: 20%) !important;
 }
 .btn-info:active,
 .btn-outline-info:active {
-  background-color: darken(#2196f3, 20%) !important;
+  background-color: color.adjust(#2196f3, $lightness: 20%) !important;
 }
 .btn-success:active,
 .btn-outline-success:active {
-  background-color: darken(#00ab55, 20%) !important;
+  background-color: color.adjust(#00ab55, $lightness: 20%) !important;
 }
 .btn-warning:active,
 .btn-outline-warning:active {
-  background-color: darken(#e2a03f, 20%) !important;
+  background-color: color.adjust(#e2a03f, $lightness: 20%) !important;
 }
 .btn-danger:active,
 .btn-outline-danger:active {
-  background-color: darken(#e7515a, 20%) !important;
+  background-color: color.adjust(#e7515a, $lightness: 20%) !important;
 }
 .btn-secondary:active,
 .btn-outline-secondary:active {
-  background-color: darken(#805dca, 20%) !important;
+  background-color: color.adjust(#805dca, $lightness: 20%) !important;
 }
 .btn-dark:active,
 .btn-outline-dark:active {
-  background-color: darken(#3b3f5c, 20%) !important;
+  background-color: color.adjust(#3b3f5c, $lightness: 20%) !important;
 }
 </style>